FirstWorkingPrototype

+++ GravitySimulation/Header.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include <GLFW/glfw3.h>
 #include <math.h>
 #define _USE_MATH_DEFINES
 #include <cmath>
 #include <vector>
+#include "glut.h"
 
-typedef unsigned long long ull;
+typedef double ull;
